refactor(client): extract file transfer & streaming into xclient_transfer.go
Move SendFile, DownloadFile, Stream out of xclient.go into
xclient_transfer.go. Close stays in xclient.go (lifecycle).
Same package, mechanical move.

xclient.go is now 231 lines (down from 1444); the God File (Critical #1)
is decomposed into 5 cohesive files: xclient.go (home/types/ctors),
xclient_discovery.go, xclient_call.go, xclient_broadcast.go,
xclient_transfer.go.
